+++ hooks/query/useReviewQuery.ts
@@ -1,11 +1,6 @@
 import { POLICY_LABEL } from "@/../interface/server/types/Auth";
 import { BeerReviewsQueryParams } from "@/../typedef/server/beer";
-import {
-  deleteReviewApi,
-  fetchAllReviewsApi,
-  getSingleReviewApi,
-  updateReviewApi,
-} from "@/api/review/api";
+
 import { dislikeReviewApi, likeReviewApi } from "@/api/review/like";
 import {
   UseMutationOptions,
@@ -22,6 +17,12 @@ import {
 } from "../../interface/types";
 import { BeerReviewsResponse } from "../../types/review/review";
 import { createReviewApi, fetchBeerReviewsApi } from "@/api/review/beer";
+import {
+  deleteReviewApi,
+  fetchAllReviewsApi,
+  getSingleReviewApi,
+  updateReviewApi,
+} from "@/api/review/review";
 
 export const allReviewsQueryKey = () => ["allReviews"];
 export const ReviewQueryKey = () => ["review"];
